requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
django-extensions==2.2.8
djangorestframework==3.11.0
ipython==7.13.0
pytest-django==3.8.0
pytest-factoryboy==2.0.3